+++ LPHK.py
@@ -38,45 +38,10 @@ else:
 VERSION = get_first_textfile_line(os.path.join(PATH, "VERSION"))
 
 # Setup dual logging/printing
-class Tee:
-    def __init__(self, file_path):
-        self.file_path = file_path
-        self.file = open(self.file_path, "w")
-        self.stdout_logger = self.TeeStdout(self.file)
-        self.stderr_logger = self.TeeStderr(self.file)
-    def __del__(self):
-        self.stdout_logger.__del__()
-        self.stderr_logger.__del__()
-        self.file.close()
-    class TeeStdout:
-        def __init__(self, file_in):
-            self.file = file_in
-            self.stdout = sys.stdout
-            sys.stdout = self
-        def __del__(self):
-            sys.stdout = self.stdout
-        def write(self, data):
-            self.file.write(data)
-            self.file.flush()
-            self.stdout.write(data)
-        def flush(self):
-            self.file.flush()
-    class TeeStderr:
-        def __init__(self, file_in):
-            self.file = file_in
-            self.stderr = sys.stderr
-            sys.stderr = self
-        def __del__(self):
-            sys.stderr = self.stderr
-        def write(self, data):
-            self.file.write(data)
-            self.file.flush()
-            self.stderr.write(data)
-        def flush(self):
-            self.file.flush()
-
 LOG_PATH = os.path.join(USER_PATH, LOG_TITLE)
-LOG = Tee(LOG_PATH)
+
+import logger
+logger.start(LOG_PATH)
 
 # Start printing output
 def datetime_str():
@@ -135,7 +100,7 @@ def shutdown():
         lp_events.timer.cancel()
         lp.Close()
         window.lp_connected = False
-    LOG.__del__()
+    logger.stop()
     if window.restart:
         if IS_EXE:
             os.startfile(sys.argv[0])
